==> files/1b_proteins.Rmd <==
---
title: "1b"
output: html_document
date: "2023-05-26"
---

```{r setup, include=FALSE} 
knitr::opts_chunk$set(warning = FALSE, message = FALSE) 
```


```{r libraries}
require("dplyr")
require("tidyr")
require("tidyverse")
require("readxl")
require("readr")
require("tibble")
require("xlsx")
require("ggplot2")
require("purrr")
require("reshape2")
require("sjmisc")

save.data = F
save.plots = F
```

```{r importing data files}
data_original <- read_tsv("data/proteinGroups.txt")
```

```{r}
data_original_plot <- data_original %>% 
  dplyr::select(ProteinID="Protein IDs", starts_with("LFQ intensity")) %>% 
  rename_all(gsub, pattern="LFQ intensity ", replacement="") %>% 
  melt(id="ProteinID") %>% 
  rename("ID" = variable, "LFQ" = value) %>% 
  mutate(LFQ = as.integer(LFQ))
```

```{r}
data_original_plot %>%
  summarize(missing_percentage = mean(is.na(LFQ) | LFQ == 0) * 100)
```

```{r}
set.seed(123)
protein_data_modified <- data_original %>% 
  dplyr::select(ProteinID="Protein IDs", starts_with("LFQ intensity")) %>% 
  dplyr::rename_all(gsub, pattern="LFQ intensity ", replacement="") %>% 
  rotate_df(cn=TRUE) %>% #, rn='ID_Nr'
  dplyr::select(where(~ any(. != 0))) 

na_data <- replace(protein_data_modified, protein_data_modified==0, NA)

```


```{r}
pMiss_row <- function(x){(rowSums(x==0)/ncol(x))*100}
pMiss_col <- function(x){(colSums(x==0)/nrow(x))*100}
```
```{r missing percentage}
#Missing percentage in each row
pMiss_row(protein_data_modified)
```

```{r}
missing.values <- na_data %>%
    gather(key = "key", value = "val") %>%
    mutate(is.missing = is.na(val)) %>%
    group_by(key, is.missing) %>%
    summarise(num.missing = n()) %>%
    filter(is.missing==T) %>%
    select(-is.missing) %>%
    arrange(desc(num.missing)) 
```

```{r}
row.plot <- na_data %>%
  mutate(id = row_number()) %>%
  gather(-id, key = "key", value = "val") %>%
  mutate(isna = is.na(val)) %>%
  ggplot(aes(key, id, fill = isna)) +
    geom_raster(alpha=0.8) +
    scale_fill_manual(name = "",
        values = c('steelblue', 'tomato3'),
        labels = c("Present", "Missing")) +
    #scale_x_discrete(limits = levels) +
    labs(x = "Protein number",
           y = "Subject ID", title = "Missing values in proteins") +
   theme(axis.text.y=element_blank(),
         axis.ticks.y=element_blank()) +
    coord_flip()

if(save.plots) ggsave(row.plot, 
                      filename = "output_files/figures/1b/missingness.png",
                      width = 10, 
                      height = 7, 
                      device='tiff', 
                      dpi=700)

```
