==> files/2a_WF1.Rmd <==
---
title: "2a"
output: html_document
date: "2023-05-26"
---

```{r setup, include=FALSE} 
knitr::opts_chunk$set(warning = FALSE, message = FALSE) 
```


```{r libraries}
require("dplyr")
require("tidyr")
require("tidyverse")
require("readxl")
require("readr")
require("tibble")
require("xlsx")
require("ggplot2")
require("purrr")
require("reshape2")
require("sjmisc")

save.data = F
save.plots = F

set.seed(123)
```

```{r importing data files}
data_perseus <- read_tsv("data/filter_30.txt")
data_baseline <- read_excel("output_files/data/1a_joined_baseline.xlsx")
```

```{r distribution plots before normalization}
# Creating modified dataset for plots
data_perseus_plot <- data_perseus[-c(1,2), ] %>% 
  dplyr::select(ProteinID="Protein IDs", 
         starts_with("LFQ intensity")) %>% 
  rename_all(gsub, pattern="LFQ intensity ", replacement="") %>% 
  melt(id="ProteinID") %>% 
  rename("ID" = variable, "LFQ" = value) %>% 
  mutate(LFQ = as.numeric(LFQ))

# Plot to check distribution of LFQ of proteins for each patient (coloured by patient-ID)
plot <- ggplot(data_perseus_plot, aes(x = LFQ, 
                                      after_stat(count), 
                                      colour = ID)) +
  geom_density() +
  labs(title = "Density plots for distribution of LFQ for each patient")+
  theme(legend.position = "none")
if(save.plots) ggsave(plot, 
                      filename = "output_files/figures/2a/1_data_distribution_patients.png",
                      width = 10, 
                      height = 7, 
                      device='tiff', 
                      dpi=700)

# Plot to check distribution of LFQ for each protein (coloured by proteins)
plot <- ggplot(data_perseus_plot, aes(x = LFQ, 
                                      after_stat(count), 
                                      colour = ProteinID)) +
  geom_density() +
  labs(title = "Density plots for distribution of LFQ for each protein")+
  theme(legend.position = "none")
if(save.plots) ggsave(plot, 
                      filename = "output_files/figures/2a/1_data_distribution_proteins.png",
                      width = 10, 
                      height = 7, 
                      device='tiff', 
                      dpi=700)
```


```{r quantile normalization}
# Quantile normalization : the aim is to give different distributions the same statistical properties
# https://ab604.github.io/docs/bspr_workshop_2018/transform.html#normalisation

quantile_normalisation <- function(df){
  
  # Find rank of values in each column
  df_rank <- map_df(df,rank,ties.method="average")
  # Sort observations in each column from lowest to highest 
  df_sorted <- map_df(df,sort)
  # Find row mean on sorted columns
  df_mean <- rowMeans(df_sorted)
  
  # Function for substituting mean values according to rank 
  index_to_mean <- function(my_index, my_mean){
    return(my_mean[my_index])
  }
  
  # Replace value in each column with mean according to rank 
  df_final <- map_df(df_rank,index_to_mean, my_mean=df_mean)
  
  return(df_final)
}

# Selecting data
data_modified <- data_perseus[-c(1,2), ] %>%                                    #excluding first 2 rows
  dplyr::select(ProteinID="Protein IDs", starts_with("LFQ intensity")) %>%             #selecting required columns
  rename_all(gsub, pattern="LFQ intensity ", replacement="")

#Quantile normalization of data
data_normalized <- data_modified %>%                                                  #selecting LFQ data
  dplyr::select(-c("ProteinID")) %>% 
  mutate_if(is.character, as.numeric) %>%                                       #converting to numeric format
  quantile_normalisation() %>%                                                  #applying quantile normalization
  bind_cols(data_modified[,1],.) %>%                                            #adding back proteinIDs
  rotate_df(cn=TRUE, rn="ID_Nr") %>%                                            #transposing dataframe
  filter(ID_Nr != 81) %>%                                                       #removing 81 and replacing it with 81_RRJ
  mutate_all(funs(str_replace(., "81_RRJ", "81"))) %>%                          # (Rosa's version)
  mutate(ID_Nr= as.integer(ID_Nr))


#data_normalized_new <- data_modified %>% 
#  #rotate_df(cn=T, rn="ID_Nr") %>% 
#  select(-ID_Nr) %>% 
#  mutate_if(is.character, as.numeric) %>%
#  quantile_normalisation() %>% 
#  bind_cols(data_modified[,1],.) %>% 
#  rename(ID_Nr = "...1") %>% 
#  filter(ID_Nr != 81) %>%                                                       #removing 81 and replacing it with 81_RRJ
#  mutate_all(funs(str_replace(., "81_RRJ", "81"))) %>%                          # (Rosa's version)
#  mutate(ID_Nr= as.integer(ID_Nr))
```


```{r plot after qnorm}
# Creating modified dataset for plots
data_perseus_plot_2 <- data_modified %>%                                                  #selecting LFQ data
  dplyr::select(-c("ProteinID")) %>% 
  mutate_if(is.character, as.numeric) %>%                                       #converting to numeric format
  quantile_normalisation()  %>%                                                  #applying quantile normalization
  bind_cols(data_modified[,1],.) %>% 
  melt(id="ProteinID") %>% 
  rename("ID" = variable, "LFQ" = value) 

# Plot to check distribution of LFQ of proteins for each patient (coloured by patient-ID)
plot <- ggplot(data_perseus_plot_2, aes(x = LFQ, 
                                      after_stat(count), 
                                      colour = ID)) +
  geom_density() +
  labs(title = "Density plots for distribution of LFQ for each patient after quantile normalization")+
  theme(legend.position = "none")
if(save.plots) ggsave(plot, 
                      filename = "output_files/figures/2a/1_data_distribution_patients_qnorm.png",
                      width = 10, 
                      height = 7, 
                      device='tiff', 
                      dpi=700)
```

```{r joining data}
joined_data <- data_baseline %>% 
  full_join(data_normalized, by = "ID_Nr") %>% 
  mutate_if(is.character, as.numeric)

if(save.data) write.xlsx(as.data.frame(joined_data), 
                         "output_files/data/2a_perseus_data.xlsx", 
                         row.names = FALSE)
```
