==> files/2b_cox.Rmd <==
---
title: "2b"
output: html_document
date: "2023-06-01"
---
```{r setup, include=FALSE} 
knitr::opts_chunk$set(warning = FALSE, message = FALSE) 
```

```{r libraries}
require("dplyr")
require("tidyr")
require("tidyverse")
require("readxl")
require("readr")
require("tibble")
require("xlsx")
require("ggplot2")
require("purrr")
require("reshape2")
require("broom")
require("gridExtra")
require("data.table")
require("survival")
require("ggfortify")
require("ggrepel")
require("survminer")
require("contsurvplot")
require("caret")
require("missForest")

save.data = F
save.plots = F

set.seed(123)
```

```{r importing data files}
annotation <- read.delim("data/proteinGroups_annotated.txt")
raw_data <- read_excel("output_files/data/2a_perseus_data.xlsx")
```

```{r creating required variables for cox}
set.seed(123)

annotation <- annotation[-1, ]                                                  #cleaning annotation file

event_data <- raw_data %>%
  mutate(event_all = if_else(is.na(Days_to_heal), 0, 1),                        #creating "event" variable
        #Days_to_heal_180 = tidyr::replace_na(Days_to_heal, 180),        
          Days_to_heal_cox = replace(Days_to_heal, 
                                     is.na(Days_to_heal),                       #replacing NA values with random value
                                     sample(seq(180, 250, 1),                   #between 180 and 250
                                            sum(is.na(Days_to_heal)), 
                                            replace = TRUE)),
         .before="A0A024R0T9;K7ER74;P02655;Q6P163;V9GYJ8") %>% 
  dplyr::select(-Days_to_heal) %>% 
  mutate(Diabetes_Type = as.factor(Diabetes_Type),
         Gender = as.factor(Gender)) 



protein_names_cox <- names(event_data[-c(1:13)])      #8                          #list of protein names

listOfModels_cox <- vector("list", length(protein_names_cox))                   #empty vector for models

# Add survival object(status == 2 is death)
event_data <- event_data %>%                                                    #survival object
  mutate(SurvObj = with(event_data, 
                        Surv(Days_to_heal_cox, event_all)),#, event_all)),
         .before="A0A024R0T9;K7ER74;P02655;Q6P163;V9GYJ8")
```

```{r boxplot for healing time}
plot <- ggplot(event_data,
       aes(x = as.factor(Healing_time),
           y = Days_to_heal_cox,
           fill = as.factor(Healing_time))) +
       geom_boxplot() +
       xlab("Healing category") + 
  ylab("Healing time") +
       labs(title = paste("Boxplots for simulated healing time categorized by Healing type")) + 
       guides(fill = "none")

if(save.plots)  ggsave(filename = paste0("output_files/figures/1a/2_boxplot_healing_simulated.png"),
                        plot,
                        width = 10, 
                        height = 7, 
                        device='tiff', 
                        dpi=700)
```


```{r barplot}
event_data$subject <- paste0("S",seq(nrow(event_data)))
event_data$status <- ifelse(event_data$event_all==1,"Event","Censored event")

plot <- ggplot(event_data, aes(subject, 
                               Days_to_heal_cox)) +
  geom_bar(stat = "identity", 
           width = 0.5, 
           position = position_dodge(0.9),
           aes(fill=as.factor(Healing_time))) +
  geom_point(data = event_data, 
             aes(subject, 
                 Days_to_heal_cox, 
                 color = as.factor(Healing_time), 
                 shape = status), 
             size = 2)  +
  coord_flip() +
  theme_bw() + 
  theme(legend.title = element_blank(),
        legend.position = "right",
        axis.text.y = element_blank()) +  
  geom_hline(yintercept = 180, 
             linetype = "dashed", 
             color = "black") +
  scale_x_discrete(expand = c(0.02, 0.02)) +
  geom_text(aes(label = "180 days", 
                y = 180, 
                x = 110),
            vjust = 0, 
            hjust = -0.1, 
            color = "black") +
  xlab("Subjects") + 
  ylab("Time to heal (days)") +
  scale_fill_discrete(labels = c("Fast-healing", "Slow-healing", "Chronic")) +
  guides(color = F)
         
if(save.plots) ggsave(plot, 
                      filename = "output_files/figures/1a/time_barplot.png",
                      width = 10, 
                      height = 7, 
                      device='tiff', 
                      dpi=700)
```

```{r cox}
results_cox = data.frame()

# loop over features
for (i in 1:length(protein_names_cox)) {
  
  fml <- paste0("SurvObj~`",protein_names_cox[i],"`+Age+Gender+Triglyceride+Diabetes_Type+Diabetes_Duration+HBA1C+BMI")
  
  currentModel <- coxph(formula = as.formula(fml), data = event_data)
  
  listOfModels_cox[[i]] <- currentModel 
  
  results_cox = rbind(results_cox, currentModel %>% tidy(conf.int = T) %>% slice(1))
  
} 
```

```{r ignore, eval=FALSE, include=FALSE}
coxph(SurvObj~CON__P00761+Age+Gender+Triglyceride+Diabetes_Type+Diabetes_Duration+HBA1C+BMI, data = event_data)

#coef — estimated hazard coefficient on the log scale
#exp(coef) — the hazard ratio (HR). Exponent of coef
#Here, 0.97 means 97%. So, CON__P00761 reduces the hazard by a factor of 3%,
# and age increases the hazard by a factor of 2.1%

coxph(SurvObj~CON__P00761+Age+Gender+Triglyceride+Diabetes_Type+Diabetes_Duration+HBA1C+BMI, data = event_data) %>% tidy(conf.int = T)
```



```{r fdr correction}
results_cox[1] <- lapply(results_cox[1], gsub, pattern = "`", replacement = "")
results_cox_adjusted <- mutate(results_cox, p.value.adjusted = p.adjust(results_cox$p.value, method = "fdr"))

results_cox_adjusted <- results_cox_adjusted %>% 
  #filter(p.value.adjusted<0.05) %>% 
  left_join(annotation, by = join_by('term' == 'Protein.IDs')) %>% 
  dplyr::select(c(term, estimate, std.error, statistic, p.value,p.value.adjusted, conf.low, conf.high, Protein.name, Protein.names, Gene.name, Gene.names, UniProt.names, PDB, KEGG, KEGG.name, GOBP, GOBP.name))

results_cox_adjusted <- results_cox_adjusted %>%
  mutate(UniProt.names = ifelse(term == "P0DOY2", "IGLC2_HUMAN", UniProt.names))
 


```

```{r saving xlsx}

if(save.data) write.xlsx(as.data.frame(results_cox_adjusted), 
                         "output_files/data/2b_cox_adjusted.xlsx", 
                         row.names = FALSE)

results_cox_filtered <- results_cox_adjusted %>% 
  filter(p.value.adjusted<0.1)

uniprot_names <- data.frame()
uniprot_names <- results_cox_filtered %>% 
  #mutate(Protein_name = str_extract(UniProt.names,"[^_]+")) %>% 
  dplyr::select(c(Gene.name, UniProt.names, estimate, std.error, statistic, p.value, p.value.adjusted, conf.low, conf.high)) %>% 
  mutate(HR = exp(estimate))

if(save.data) write.xlsx(as.data.frame(uniprot_names), 
                         "output_files/data/2b_cox_adjusted_uniprotnames.xlsx", 
                         row.names = FALSE)
  
```



```{r corrplot}
require("corrplot")

# Correlation plot for proteins vs proteins

corplot_list <- results_cox_filtered$term
corplot_gene_names <- str_extract(results_cox_filtered$Gene.name,"[^;]+")

corplot_data <- raw_data[, corplot_list] %>% 
  set_names(corplot_gene_names)

M<-cor(corplot_data, method = "spearman")

png("output_files/figures/2b/2b_corr_proteins.png", width=7.5, height=7.5, units = "in", res = 300)
gplots::heatmap.2(M,
                  revC = TRUE,
                  trace = "none",
                  density.info = "none",
                  #margins = c(8, 8),
                  cexRow = 0.9,
                  cexCol = 0.9)
dev.off()
```
```{r}
# Correlation plot for proteins vs covariates

covariates <- c("Age","Gender", "Triglyceride", "Diabetes_Type", "Diabetes_Duration", "HBA1C", "BMI")
corplot_data <- raw_data[, corplot_list] %>% 
  set_names(corplot_gene_names) %>% 
  cbind(raw_data[, covariates])
  
M<-cor(corplot_data, use='pairwise.complete.obs', method = "spearman")


M_fin <- M[covariates,corplot_gene_names]

png("output_files/figures/2b/2b_corr_covariates.png", width=7.5, height=7.5, units = "in", res = 300)
gplots::heatmap.2(M_fin, #Colv = NA, 
                  Rowv = NA, 
                  revC = TRUE,
    trace = "none", 
    density.info = "none",
    margins = c(6, 8),
    cexRow = 0.9,
    cexCol = 0.9)
dev.off()
```


```{r volcano plot}
# https://ab604.github.io/docs/bspr_workshop_2018/transform.html#volcano-plot

plot <- ggplot(results_cox_adjusted, aes(x = exp(estimate), y = -log10(p.value.adjusted))) +
  geom_point(aes(color="grey", 
                 alpha=0.4,
                 size = -log10(p.value.adjusted)))  +
  #ggtitle("Volcano plot for cox survival analysis estimates") +
  xlab("Hazard ratio") + 
  ylab(expression("-log"[10]*"(p-value)")) + 
  theme_bw() + 
  theme(plot.title = element_text(size = rel(1.5), 
                                  hjust = 0, 
                                  margin = margin(b = 10))) +
    geom_text_repel(data=results_cox_filtered,
                    aes(x = exp(estimate), 
                        y = -log10(p.value.adjusted),
                        label=str_extract(Gene.name,"[^;]+"),  # (UniProt.names,"[^_]+")
                        point.size = -log10(p.value.adjusted)),
                    size=2.5, 
                    point.padding = 0,
                    #min.segment.length = 0,
                    max.time = 3, max.iter = 1e6,
                    box.padding = 0.3,
                    max.overlaps = 20) +
  geom_point(data=results_cox_filtered,
             aes(x = exp(estimate), 
                 y = -log10(p.value.adjusted),
                 color="red",
                 alpha=0.4,
                 size = -log10(p.value.adjusted))) +
  scale_color_manual(values=c('gray','red')) + 
  geom_vline(xintercept = 1, linetype = "dotted")+ 
  #xlim(c(-.5, .5)) + 
         scale_color_manual(labels = c("None", expression("P"[FDR]*"< 0.10")), 
                     values = c("grey", "red"))+ 
  labs(color = "Significance") +
  guides(color = guide_legend(title = "Significance"),
         alpha = FALSE,  # Remove the alpha legend
         size = FALSE)   # Remove the size legend

if(save.plots) ggsave(plot, 
                      filename = "output_files/figures/2b/volcano_plot.png",
                      width = 10, 
                      height = 7, 
                      device='tiff', 
                      dpi=100)
```

```{r KM survival curves - covariates}
# https://en.wikipedia.org/wiki/Survival_analysis
# https://rviews.rstudio.com/2017/09/25/survival-analysis-with-r/

# add method to grid.draw
grid.draw.ggsurvplot <- function(x){
  survminer:::print.ggsurvplot(x, newpage = FALSE)
}

# All

fit <- survfit(event_data$SurvObj ~ 1, data=event_data) 

if(save.data) sink("output_files/figures/2b/fit_summary_all.txt")
print(summary(fit, times = c(1,5*(1:40))))
cat("\n")
print(coxph(Surv(Days_to_heal_cox) ~ 1, data=event_data))
sink()

plot <- ggsurvplot(fit,
                   fun = 'event',
                   conf.int=TRUE, 
                   pval=TRUE,
                   pval.coord=c(200,0.2), 
                   risk.table=TRUE,
                   cumevents=TRUE, 
           #legend.labs=c("Type 1", "Type 2"), 
           #legend.title="Diabetes type",  
           palette=c("dodgerblue2", "orchid2"), 
           title="Kaplan-Meier Curve for DFU healing", 
           xlab="Time in days",
           ylab="Probability of healing (%)",
           risk.table.height=.15,
           risk.table.title="Number of potentially healing subjects",
           cumevents.height=.15,
           cumevents.title="Cumulative number of healed subjects")

if(save.plots) ggsave(plot, 
                      filename = "output_files/figures/2b/KM_curve_all.png",
                      width = 7.5,
                      height = 7.5)


# Diabetes type

fit <- survfit(event_data$SurvObj ~ Diabetes_Type, data=event_data) 

if(save.data) sink("output_files/figures/2b/fit_summary_Dtype.txt")
print(summary(fit, times = c(1,5*(1:40))))
cat("\n")
print(coxph(Surv(Days_to_heal_cox) ~ Diabetes_Type, data=event_data))
sink()

plot <- ggsurvplot(fit,
                   fun = 'event',
                   conf.int=TRUE, 
                   pval=TRUE,
                   pval.coord=c(200,0.2),
                   risk.table=TRUE,
                   cumevents=TRUE,
           legend.labs=c("Type 1", "Type 2"), 
           legend.title="Diabetes type",  
           palette=c("dodgerblue2", "orchid2"), 
           title="Kaplan-Meier Curve for DFU healing [Diabetes Type]",
           xlab="Time in days",
           ylab="Probability of healing (%)",
           risk.table.height=.15,
           risk.table.title="Number of potentially healing subjects",
           cumevents.height=.15,
           cumevents.title="Cumulative number of healed subjects")

if(save.plots) ggsave(file = "output_files/figures/2b/KM_curve_Dtype.png",
                      plot,
                      width = 7.5,
                      height = 7.5)

# Gender

fit <- survfit(event_data$SurvObj ~ Gender, data=event_data) 

if(save.data) sink("output_files/figures/2b/fit_summary_gender.txt")
print(summary(fit, times = c(1,5*(1:40))))
cat("\n")
print(coxph(Surv(Days_to_heal_cox) ~ Gender, data=event_data))
sink()

plot <- ggsurvplot(fit,
                   fun = 'event',
                   conf.int=TRUE, 
                   pval=TRUE,
                   pval.coord=c(200,0.2),
                   risk.table=TRUE,
                   cumevents=TRUE,
           legend.labs=c("Male", "Female"), 
           legend.title="Gender",  
           palette=c("dodgerblue2", "orchid2"), 
           title="Kaplan-Meier Curve for DFU healing [Gender]",
           xlab="Time in days",
           ylab="Probability of healing (%)",
           risk.table.height=.15,
           risk.table.title="Number of potentially healing subjects",
           cumevents.height=.15,
           cumevents.title="Cumulative number of healed subjects")

if(save.plots) ggsave(file = "output_files/figures/2b/KM_curve_gender.png",
                      plot,
                      width = 7.5,
                      height = 7.5)
```

```{r KM survival contours - covariates}
# Diabetes duration

fit <- coxph(Surv(Days_to_heal_cox) ~ Diabetes_Duration, data=event_data, x=TRUE)
pvalue <- round(summary(fit)$coefficients[, 5], digits = 2)

if(save.data) sink("output_files/figures/2b/fit_summary_Dduration.txt")
print(survfit(event_data$SurvObj ~ Diabetes_Duration, data=event_data))
cat("\n")
print(coxph(Surv(Days_to_heal_cox) ~ Diabetes_Duration, data=event_data))
sink()

plot <- plot_surv_contour(time="Days_to_heal_cox",
                  status="event_all",
               variable="Diabetes_Duration",
               data=event_data,
               model=fit,
               fun = "event",
               title="Survival Probability Contour Curve for DFU Healing [Diabetes Duration]",
               legend.title="Probability of not healing",
               xlab="Time (days)",
               ylab="Diabetes duration (years)") + 
  annotate("text", x = -Inf, y = -Inf,
           vjust = -2, hjust = -1, 
           label = paste0("p = ",pvalue))

if(save.plots)ggsave(file = "output_files/figures/2b/KM_curve_Dduration.png",
                      plot,
                      width = 7.5,
                      height = 4.5)

# BMI

fit <- coxph(Surv(Days_to_heal_cox) ~ BMI, data=event_data, x=TRUE)
pvalue <- round(summary(fit)$coefficients[, 5], digits = 2)

if(save.data) sink("output_files/figures/2b/fit_summary_bmi.txt")
print(survfit(event_data$SurvObj ~ BMI, data=event_data))
cat("\n")
print(coxph(Surv(Days_to_heal_cox) ~ BMI, data=event_data))
sink()

plot <- plot_surv_contour(time="Days_to_heal_cox",
                  status="event_all",
               variable="BMI",
               data=event_data,
               model=fit,
               title="Survival Probability Contour Curve for DFU Healing [BMI]",
               legend.title="Probability of not healing",
               xlab="Time (days)",
               ylab="Body Mass Index (BMI)") + 
  annotate("text", x = -Inf, y = -Inf,
           vjust = -2, hjust = -1, 
           label = paste0("p = ",pvalue))

if(save.plots) ggsave(file = "output_files/figures/2b/KM_curve_bmi.png",
                      plot,
                      width = 7.5,
                      height = 4.5)

# age

fit <- coxph(Surv(Days_to_heal_cox) ~ Age, data=event_data, x=TRUE)
pvalue <- round(summary(fit)$coefficients[, 5], digits = 2)

if(save.data) sink("output_files/figures/2b/fit_summary_age.txt")
print(survfit(event_data$SurvObj ~ Age, data=event_data))
cat("\n")
print(coxph(Surv(Days_to_heal_cox) ~ Age, data=event_data))
sink()

plot <- plot_surv_contour(time="Days_to_heal_cox",
                  status="event_all",
               variable="Age",
               data=event_data,
               model=fit,
               title="Survival Probability Contour Curve for DFU Healing [Age]",
               legend.title="Probability of not healing",
               xlab="Time (days)",
               ylab="Age (years)") + 
  annotate("text", x = -Inf, y = -Inf,
           vjust = -2, hjust = -1, 
           label = paste0("p = ",pvalue))

if(save.plots) ggsave(file = "output_files/figures/2b/KM_curve_age.png",
                      plot,
                     width = 7.5,
                      height = 4.5)

# hba1c

fit <- coxph(Surv(Days_to_heal_cox) ~ HBA1C, data=event_data, x=TRUE)
pvalue <- round(summary(fit)$coefficients[, 5], digits = 2)

if(save.data) sink("output_files/figures/2b/fit_summary_HBA1C.txt")
print(survfit(event_data$SurvObj ~ HBA1C, data=event_data))
cat("\n")
print(coxph(Surv(Days_to_heal_cox) ~ HBA1C, data=event_data))
sink()

plot <- plot_surv_contour(time="Days_to_heal_cox",
                  status="event_all",
               variable="HBA1C",
               data=event_data,
               model=fit,
               title="Survival Probability Contour Curve for DFU Healing [HBA1C]",
               legend.title="Probability of not healing",
               xlab="Time (days)",
               ylab="HBA1C measure (mmol/mol)") + 
  annotate("text", x = -Inf, y = -Inf,
           vjust = -2, hjust = -1, 
           label = paste0("p = ",pvalue))

if(save.plots) ggsave(file = "output_files/figures/2b/KM_curve_HBA1C.png",
                      plot,
                      width = 7.5,
                      height = 4.5)

# Triglycerides

fit <- coxph(Surv(Days_to_heal_cox) ~ Triglyceride, data=event_data, x=TRUE)
pvalue <- round(summary(fit)$coefficients[, 5], digits = 2)

if(save.data) sink("output_files/figures/2b/fit_summary_triglycerides.txt")
print(survfit(event_data$SurvObj ~ Triglyceride, data=event_data))
cat("\n")
print(coxph(Surv(Days_to_heal_cox) ~ Triglyceride, data=event_data))
sink()

plot <- plot_surv_contour(time="Days_to_heal_cox",
                  status="event_all",
               variable="Triglyceride",
               data=event_data,
               model=fit,
               title="Survival Probability Contour Curve for DFU Healing [Triglycerides]",
               legend.title="Probability of not healing",
               xlab="Time (days)",
               ylab="Triglyceride measure (mmol/l)") + 
  annotate("text", x = -Inf, y = -Inf,
           vjust = -2, hjust = -1, 
           label = paste0("p = ",pvalue))

if(save.plots) ggsave(file = "output_files/figures/2b/KM_curve_triglycerides.png",
                      plot,
                      width = 7.5,
                      height = 4.5)
```

```{r KM survival contours - proteins}



for (i in 1:nrow(results_cox_filtered)) {
  
  fml <- paste0("SurvObj~`",results_cox_filtered$term[i],"`+Age+Gender+Triglyceride+Diabetes_Type+Diabetes_Duration+HBA1C+BMI")
  
  fit <- coxph(formula = as.formula(fml), data=event_data, x=TRUE)
  pvalue <- round(results_cox_filtered$p.value.adjusted[i], digits = 3)
  
  
  if(save.data) sink(paste0("output_files/figures/2b/fit_summary_", 
                                      str_extract(results_cox_filtered$Gene.name[i],"[^;]+"), ".txt"))
  print(summary(fit))
  sink()
  
  
  plot <- plot_surv_contour(time="Days_to_heal_cox",
                  status="event_all",
               variable=results_cox_filtered$term[i],
               data=event_data,
               model=fit,
               title=paste0("Survival Probability Contour Curve for DFU Healing [Protein -> ", 
                            str_extract(results_cox_filtered$Gene.name[i],"[^;]+"), "]"),
               legend.title="Probability of not healing",
               xlab = "Time (days)",
               ylab = str_extract(results_cox_filtered$Gene.name[i],"[^;]+")) + 
  annotate("text", x = -Inf, y = -Inf,
           vjust = -2, hjust = -1, 
           label = bquote("P"[FDR]*"="~ .(pvalue)),
           colour = "white")
  
  if(save.plots) ggsave(file = paste0("output_files/figures/2b/2b_KM_curve_", 
                                      str_extract(results_cox_filtered$Gene.name[i],"[^;]+"), ".png"),
                                      plot,
                                      width = 7.5,
                        height = 4.5)
  
} 

```
