==> files/4a_WF3.Rmd <==
---
title: "4a"
output: html_document
date: "2023-06-22"
---

```{r setup, include=FALSE} 
knitr::opts_chunk$set(warning = FALSE, message = FALSE) 
```


```{r libraries}
require("dplyr")
require("tidyr")
require("tidyverse")
require("readxl")
require("readr")
require("tibble")
require("xlsx")
require("ggplot2")
require("purrr")
require("reshape2")
require("sjmisc")
require("mice")
require("robCompositions")

save.data = F
save.plots = F

set.seed(123)
```

```{r importing data files}
data_proteins <- read_tsv("data/proteinGroups.txt")
data_baseline <- read_excel("output_files/data/1a_joined_baseline.xlsx")
```

```{r}
set.seed(123)
protein_data_modified <- data_proteins %>% 
  dplyr::select(ProteinID="Protein IDs", starts_with("LFQ intensity")) %>% 
  dplyr::rename_all(gsub, pattern="LFQ intensity ", replacement="") %>% 
  rotate_df(cn=TRUE) %>% #, rn='ID_Nr'
  dplyr::select(where(~ any(. != 0))) 

na_data <- replace(protein_data_modified, protein_data_modified==0, NA)

pattern <- "protein_%d"

# Create the new names using sprintf
new_names <- sprintf(pattern, 1:ncol(na_data))



# Create the dictionary
name_dict <- setNames(colnames(na_data), new_names)

names(na_data) <- new_names

```




```{r}
pMiss_row <- function(x){(rowSums(x==0)/ncol(x))*100}
pMiss_col <- function(x){(colSums(x==0)/nrow(x))*100}
```
```{r}
pMiss_row(protein_data_modified)
```

```{r}
missing.values <- na_data %>%
    gather(key = "key", value = "val") %>%
    mutate(is.missing = is.na(val)) %>%
    group_by(key, is.missing) %>%
    summarise(num.missing = n()) %>%
    filter(is.missing==T) %>%
    select(-is.missing) %>%
    arrange(desc(num.missing)) 
```


```{r}
tempData <- mice(na_data,method = 'rf', seed=123, parallel=8)
```

```{r}
completedData_1 <- complete(tempData,1)
completedData_2 <- complete(tempData,2)
completedData_3 <- complete(tempData,3)
completedData_4 <- complete(tempData,4)
completedData_5 <- complete(tempData,5)
```

```{r}
completedData_final <- (completedData_1+completedData_2+completedData_3+completedData_4+completedData_5)/5
```

```{r}
#to replace names back
colnames(completedData_final) <- name_dict
```

```{r}
if(save.data) write.xlsx(as.data.frame(completedData_final), 
                         "output_files/data/4a_mice_proteins_data.xlsx", 
                         row.names = T)
```

```{r}
completed_data_modified <- rownames_to_column(completedData_final, "ID_Nr") %>%
  filter(ID_Nr != 81) %>%                                                       #removing 81 and replacing it with 81_RRJ
  mutate_all(funs(str_replace(., "81_RRJ", "81"))) %>%                          # (Rosa's version)
  mutate(ID_Nr= as.integer(ID_Nr))
```


```{r filter 50%}
# Calculate the proportion of non-zero values for each protein
protein_counts <- completed_data_modified %>%
  select(-ID_Nr) %>% 
  summarise(across(everything(), ~ mean(is.na(.), na.rm = TRUE)))

# Identify proteins present in at least 50% of the subjects
selected_proteins <- protein_counts %>%
  select(where(~ mean(.) <= 0.7)) %>%
  colnames()

# Retain only the selected proteins in the data
filtered_data <- completed_data_modified %>%
  select(ID_Nr, all_of(selected_proteins))

log_transformed <- filtered_data %>%
  select(-ID_Nr) %>% 
  #na.omit() %>% 
 mutate_if(is.character, as.numeric) %>% 
  cenLR()
  #log2() %>% 
  #rownames_to_column(var = "ID_Nr") %>% 
  #rotate_df(cn=TRUE, rn="ProteinID")

log_transformed_id <- cbind(ID_Nr = filtered_data$ID_Nr, log_transformed$x.clr)

#qnorm_data <- log_transformed %>%
#  dplyr::select(-c("ProteinID")) %>% 
#  mutate_if(is.character, as.numeric) %>%                                       
#  quantile_normalisation()  %>%                                                  
#  bind_cols(log_transformed[,1],.) %>% 
#  rename(ProteinID = "...1") %>%                                            #adding back proteinIDs
#  rotate_df(cn=TRUE, rn="ID_Nr") %>% 
#  mutate_if(is.character, as.numeric)
  
```


```{r}
joined_data <- data_baseline %>% 
  inner_join(log_transformed_id, by = "ID_Nr") %>% 
  mutate_if(is.character, as.numeric)

if(save.data) write.xlsx(as.data.frame(joined_data), 
                         "output_files/data/4a_mice_data.xlsx", 
                         row.names = FALSE)
```

```{r}
saveRDS(tempData, file = "output_files/data/4a_mice_mids.Rdata")
```
